Refactor database schema and views for consistency and improved data types
- Changed data type of "dividends" from REAL to DOUBLE PRECISION in StockPricesYahooHistoryWeekly and StockPricesYahooMasterData tables.
- Updated various FLOAT columns to DOUBLE PRECISION in TechnicalIndicatorsHistoryDaily, TechnicalIndicatorsHistoryMonthly, TechnicalIndicatorsHistoryWeekly, and TechnicalIndicatorsMasterData tables for better precision.
- Created a new view StockAssetProfilesYahooHistory to consolidate asset profile data from master and daily tables.
- Modified database.py to remove SQLite support and enhance column retrieval logic for PostgreSQL.
- Updated historization.py to utilize the new get_columns function for fetching column details.

+++ db/SQL/tables/create_table/history/AnalystPriceTargetsHistoryDaily.sql
@@ -3,7 +3,7 @@
     CREATE TABLE IF NOT EXISTS "AnalystPriceTargetsHistoryDaily" (
         snapshot_date DATE NOT NULL,
         "symbol" TEXT,
-        "analyst_mean_target" FLOAT,
+        "analyst_mean_target" DOUBLE PRECISION,
         PRIMARY KEY(snapshot_date, "symbol")
     );
